Don't call KDirWatch for remote urls, fixes a runtime warning when reloading a FTP url.
svn path=/branches/KDE_3_4_BRANCH/kdelibs/; revision=405186

+++ kio/kio/kdirlister_p.h
@@ -239,7 +239,7 @@ private:
     {
       if ( autoUpdates )
       {
-        if ( KDirWatch::exists() )
+        if ( KDirWatch::exists() && url.isLocalFile() )
           kdirwatch->removeDir( url.path() );
         sendSignal( false, url );
       }
